feat(values): implements get real values, adjusts visuals like width
