+++ src/modules/vault/components/bridge/amount.tsx
@@ -1,55 +1,25 @@
 import { Button, Card, HStack, Text, VStack } from '@chakra-ui/react';
 import { motion } from 'framer-motion';
-import { useMemo, useState } from 'react';
-import { useFormContext, useWatch } from 'react-hook-form';
+import { useState } from 'react';
 
-import { ITransferBridgePayload } from '../../hooks/bridge';
 import { InputAmount } from './inputAmount';
 
 export interface AmountBridgeProps {
+  symbol: string;
   stepsForm: number;
   setStepsForm: React.Dispatch<React.SetStateAction<number>>;
 }
 
 const MotionBox = motion(VStack);
 
-const optionsAssets = [
-  {
-    value: '0xf8f8b6283d7fa5b672b530cbb84fcccb4ff8dc40f8176ef4544ddb1f1952ad07',
-    name: 'ETH',
-    image: 'https://assets.fuel.network/providers/eth.svg',
-    symbol: 'ETH',
-  },
-  {
-    value: '0x1d5d97005e41cae2187a895fd8eab0506111e0e2f3331cd3912c15c24e3c1d82',
-    name: 'FUEL',
-    image: 'https://verified-assets.fuel.network/images/fuel.svg',
-    symbol: 'FUEL',
-  },
-  {
-    value: 'USDC',
-    name: 'USDC',
-    image:
-      'https://firebasestorage.googleapis.com/v0/b/pump-555ee.appspot.com/o/images%2Faecb0358-d860-402c-9f3c-c5b579e4eb88.jpeg?alt=media&token=b39c9a29-4b5e-4b2c-8600-62e9afff2448',
-    symbol: 'USDC',
-  },
-];
-
-export function AmountBrigde({ stepsForm, setStepsForm }: AmountBridgeProps) {
-  const { control } = useFormContext<ITransferBridgePayload>();
+export function AmountBrigde({
+  symbol,
+  stepsForm,
+  setStepsForm,
+}: AmountBridgeProps) {
+  //const { control } = useFormContext<ITransferBridgePayload>();
   const [valueSource, setValueSource] = useState('0.000');
   const [errorAmount, setErrorAmount] = useState(false);
-
-  const assetFromValue = useWatch({
-    control,
-    name: 'selectAssetFrom',
-  });
-
-  const symbol =
-    useMemo(
-      () => optionsAssets.find((a) => a.value === assetFromValue)?.symbol,
-      [assetFromValue],
-    ) ?? '';
 
   const handleSourceChange = (value: string) => {
     console.log('value', value);
